+++ src/main/groovy/org/codenarc/rule/naming/FieldNameRule.groovy
@@ -104,7 +104,7 @@ class FieldNameAstVisitor extends AbstractAstVisitor  {
             }
 
             if (!(fieldNode.name ==~ re)) {
-                addViolation(fieldNode, "The fieldname $fieldNode.name in class ${fieldNode.owner?.name} does not match ${re.toString()}")
+                addViolation(fieldNode, "The fieldname $fieldNode.name in class ${fieldNode.owner?.name} does not match ${re}")
             }
         }
         super.visitField(fieldNode)