+++ src/main/groovy/org/codenarc/rule/naming/VariableNameRule.groovy
@@ -61,7 +61,7 @@ class VariableNameAstVisitor extends AbstractAstVisitor  {
             varExpressions.each { varExpression ->
                 if (!new WildcardPattern(rule.ignoreVariableNames, false).matches(varExpression.name) &&
                         !(varExpression.name ==~ re)) {
-                    def msg = "Variable named $varExpression.name in class $currentClassName does not match the pattern ${re.toString()}"
+                    def msg = "Variable named $varExpression.name in class $currentClassName does not match the pattern ${re}"
                     addViolation(declarationExpression, msg)
                 }
             }