+++ src/main/groovy/org/codenarc/source/CustomCompilerPhaseSourceDecorator.groovy
@@ -65,6 +65,6 @@ class CustomCompilerPhaseSourceDecorator extends AbstractSourceCode {
 
     @Override
     String toString() {
-        "CustomCompilerPhaseSourceDecorator[${delegate.toString()}]"
+        "CustomCompilerPhaseSourceDecorator[${delegate}]"
     }
 }